# labs/lab03-rest-api/starter/openapi.yaml
openapi: 3.0.3
info:
  title: Lab 3 Item API
  version: 1.0.0
  description: TODO - briefly describe this API.

servers:
  - url: http://localhost:3000

paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

  /items:
    get:
      summary: Get all items
      responses:
        "200":
          description: Get all items in the server
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      summary: Create item
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Item'
        "400":
          description: Error with user input
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Item does not exist
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'

      requestBody:
        description: Item data to send
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Item"
  /items/{id}:
    summary: Find, update, or delete an item
    parameters:
      - in: query
        name: id
        schema:
          type: number
    get:
      responses:
        "200":
          description: "Get an item"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Item'
        "404":
          description: Item does not exist
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'

    put:
      responses:
        "200":
          description: "Update an item"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/Item'
        "404":
          description: Item does not exist
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        description: Item data to send
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
    delete:
      responses:
        "204":
          description: "Delete an item"
        "404":
          description: "Item does not exist"
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/ErrorResponse'



  #
  # Remember:
  # - Express routes use :id, such as /items/:id.
  # - OpenAPI paths use {id}, such as /items/{id}.
  #
  # You may define schemas under components/schemas if you want to avoid
  # repeating the same object structure many times.

components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
          example: 2
          format: uint
        name:
          example: "clock"
          type: string
        quantity:
          example: 6
          type: integer
          format: uint
      required:
        - id
        - name
        - quantity
    ItemInput:
      type: object
      properties:
        name:
          example: "clock"
          type: string
        quantity:
          example: 6
          type: integer
          format: uint
      required:
        - name
        - quantity
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Item does not exist"
      required:
        - error